Update wine.eselect module Q/A fix handling of symbolic links created to point at variant files with file extensions. Move to a consistent approach of putting variant name _before_ any file extension in all cases. E.g.:
<file-base>[-variant][.<file-extension>]

The wrapper files, created by the app-emulation/wine-{staging,vanilla}
packages, are now updated to be consistent with this format.

Allow specifying app-emulation/wine-{staging,vanilla} targets as:
  ${PN}:${PV} (new style - compatible with Portage)
or
  ${PN}-${PV} (old style)
All targets are now displayed as:
  ${PN}:${PV}
when using the eselect wine show|list options.

Module version updated to: 1.5.3

The manpage for eselect.wine has been updated to reflect the (above) changes.